backend/pyproject.toml
[tool.poetry]
name = "shinkei-backend"
version = "0.1.0"
description = "Shinkei Narrative Engine - Backend API"
authors = ["Shinkei Team <team@shinkei.dev>"]
packages = [{include = "shinkei", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.17"
httpx = "^0.28.0"
supabase = "^2.9.1"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"
opentelemetry-instrumentation-fastapi = "^0.49b2"
structlog = "^24.4.0"
email-validator = "^2.3.0"
openai = "^2.8.1"
anthropic = "^0.74.0"
ollama = "^0.6.1"
slowapi = "^0.1.9"
bleach = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
black = "^24.10.0"
ruff = "^0.7.4"
mypy = "^1.13.0"
bandit = "^1.7.5"
safety = "^3.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
